feat: add nav command handlers
