+++ src/app.js
@@ -1,12 +1,40 @@
+import { createInterface } from "node:readline";
+import { stdin, stdout, chdir } from "node:process";
+import { homedir } from "node:os";
 
-import CommandController from "./controllers/index.js";
+import CommandExecutor from "./commandExecutor.js";
+import { commands } from "./commands/commandDescriptors.js";
+import parseCommand from "./utils/parseCommand.js";
+import getCommandMetadata from "./utils/getCommandMetadata.js";
 
 class App {
-  #commandController = new CommandController();
+  #commandExecutor = new CommandExecutor();
 
-  start() {
-    this.#commandController.run();
+  run() {
+    chdir(homedir());
+
+    const readLine = createInterface({
+      input: stdin,
+      output: stdout,
+    });
+
+    this.#commandExecutor.exec(commands.greeting);
+
+    readLine
+      .on("line", (input) => {
+        const parsedCommand = parseCommand(input);
+        const { isValid, commandKey } = getCommandMetadata(parsedCommand);
+
+        if (isValid) {
+          this.#commandExecutor.exec(commandKey, ...parsedCommand.arguments);
+        } else {
+          this.#commandExecutor.exec(commands.invalidCommand);
+        }
+      })
+      .on("close", () => {
+        this.#commandExecutor.exec(commands.exit);
+      });
   }
 }
 
-export default new App();
+export default App;
